fossup.py: Initial Commit.
This is the inital commit of fossup.py. It is used to upload file/
archives to a fossology instance.

It is tested with both python 2 and 3. It is tested with
Fossology 3.2.0rc1-20-gce3d8bc (code revision ce3d8b)
and is shown to work.

General useage is:

fossup.py -n "My Test Directory" -f archive.tgz -d "Test Archive"

See README and fossup.py -h for more info

Signed-off-by: Eilís Ní Fhlannagáin <pidge@toganlabs.com>